requirements.txt
aioconsole==0.8.2
alembic==1.18.4
asyncpg==0.31.0
bcrypt==5.0.0
fastapi==0.135.3
httpx==0.28.1
pydantic==2.12.5
pydantic_settings==2.13.1
pytest==9.0.2
pytest_asyncio==1.3.0
python_bcrypt==0.3.2
python_jose==3.5.0
redis==7.4.0
Requests==2.33.1
SQLAlchemy==2.0.49
uvicorn==0.41.0
websockets==16.0
